requirements.txt
streamlit
PyGithub==1.55
bcrypt
PyJWT>=2.0.0  
python-jose>=3.3.0
pandas
python-dotenv
plotly==5.5.0
kaleido==0.2.1
streamlit_cookies_controller
reportlab